Add files via upload
Look at the try # 3. If we can transform the temp file into a gtiff file. It will make no. of bands = time stamps, and it will become easier to extract output for each date. I am also looking for a way to clip/mask the tiff to a specific polygon/multipolygon